add unit tests and address comments
