+++ modules/layers/src/text-layer/font-atlas-manager.js
@@ -4,7 +4,7 @@ import {Texture2D} from '@luma.gl/core';
 import TinySDF from '@mapbox/tiny-sdf';
 import GL from '@luma.gl/constants';
 
-import {buildMapping} from './font-atlas-utils';
+import {buildMapping} from './utils';
 import LRUCache from './lru-cache';
 
 function getDefaultCharacterSet() {
