+++ test/modules/layers/index.js
@@ -29,7 +29,7 @@ import './simple-mesh-layer.spec';
 import './scenegraph-layer.spec';
 import './path-layer/path-layer-vertex.spec';
 import './icon-manager.spec';
-import './text-layer/font-atlas-utils.spec';
+import './text-layer/utils.spec';
 import './text-layer/lru-cache.spec';
 import './text-layer/text-layer.spec';
 import './column-geometry.spec';
